Changes preprocessing function
I have reason to believe the preprocessing function is incorrect. It seems to be ignoring the channels in the image and the results are off. With reference to the onnx.js implementation of the same model [here](https://github.com/Microsoft/onnxjs-demo/blob/master/src/components/models/Emotion.vue#L48), they seem to be first grayscaling the image if indeed it has 3/4 channels, then they are centering the values around 0 by subtracting 127.5 and dividing by the same. I am proposing to change the preprocessing function. Currently written for clarity, feel free to shorten. I am working on an [implementation](http://app.modelhub.ai/) and found that the onnx.js preprocessing gives more realistic results (they used the HTML canvas to do the resizing to 64x64, so there is really no concrete way to fully reproduce their implementation outside www and benchmark against it).