Remove unnecessary reshape in mnist-1 model
This is how the model was changed:
```
-- model.prototxt.orig 2020-12-22 12:16:40.634961643 +0100
+++ model.prototxt      2020-12-22 12:18:40.589855609 +0100
@@ -6409,21 +6409,8 @@
     domain: ""
   }
   node {
-    input: "Constant367"
-    output: "Reshape420_Output_0"
-    name: "Reshape420"
-    op_type: "Reshape"
-    attribute {
-      name: "shape"
-      ints: 10
-      type: INTS
-    }
-    doc_string: ""
-    domain: ""
-  }
-  node {
     input: "Times418_Output_0"
-    input: "Reshape420_Output_0"
+    input: "Constant367"
     output: "Plus422_Output_0"
     name: "Plus214"
     op_type: "Add"
```

The unnecessary node did reshape of Constant367 (which has shape [1, 10]) to shape [10],
and the output of that operation was passed as a second input to Add node (named "Plus214").
Since first input to "Plus214" node has shape [1, 10], reshape described above
made the model incorrect because input shapes to Add operator didn't match and its broadcast attribute is set to 0.

Signed-off-by: Mateusz Tabaka <mateusz.tabaka@intel.com>